chasm/lib/activity: keep transition bodies with their transitions
The previous commit scattered the 15 transition bodies across attempt.go,
outcome.go and operator_commands.go, placing each by topic. That was the
wrong axis. Every one of them has exactly one caller - its own transition
in statemachine.go - and exists solely to be that transition's body, so
the caller is the content: applyCompleted's subject is "the body of
TransitionCompleted", not "outcomes".

Placing them by topic also meant the topics had to be argued, and two of
the three files ended up with a name that did not describe their
contents: outcome.go held applyCompleted/Failed/Terminated/Canceled but
not applyTimedOut, which is equally a terminal outcome and was filed
under attempt.go because it switches on timeout type.

Move all 15 into statemachine.go, below the state table, in the same
order as the transitions above them. The rule is now structural rather
than topical, and mechanically checkable: a transition body lives beneath
its transition and is called from exactly nowhere else.

  statemachine.go        337 -> 788
  operator_commands.go   749 -> 597
  attempt.go             610 -> 458
  outcome.go             deleted

outcome.go had only ever held the four terminal bodies plus
recordScheduleToStartOrCloseTimeoutFailure, which moves back to
attempt.go beside recordFailedAttempt, the other failure-recording
helper. Both are called only from applyTimedOut.

statemachine.go is now the second-largest file in the package, but it
reads as two clearly separated halves: an 18-transition state table that
fits on a couple of screens, then the named bodies. Before this series it
was 652 lines of state table interleaved with anonymous closures.

Three apply*-named methods stay where they are - applyFailedAttempt,
applyDeferredOptionRestore and applyDeferredHeartbeatClear are shared
helpers with two callers each, not transition bodies. That the prefix
means two things is worth fixing separately.

Pure move, verified with .task/ast-diff: all 335 declarations plus their
doc comments and the 72-entry import union are BYTE-IDENTICAL to the
parent commit. The only comment changes are outcome.go's file-scope
comment (removed with the file) and the rewording of the group comment
above the bodies.

Run the tests from the repo root with:
  export TEMPORAL_TEST_LOG_LEVEL=ERROR TEMPORAL_TEST_LOG_STACKTRACE_LEVEL=off
  go test ./chasm/lib/activity/...
