+++ chasm/lib/activity/statemachine.go
@@ -1,16 +1,21 @@
 package activity
 
 import (
+	"fmt"
 	"time"
 
 	commonpb "go.temporal.io/api/common/v1"
+	deploymentpb "go.temporal.io/api/deployment/v1"
 	enumspb "go.temporal.io/api/enums/v1"
 	failurepb "go.temporal.io/api/failure/v1"
 	"go.temporal.io/api/workflowservice/v1"
 	"go.temporal.io/server/api/historyservice/v1"
 	"go.temporal.io/server/chasm"
 	"go.temporal.io/server/chasm/lib/activity/gen/activitypb/v1"
+	"go.temporal.io/server/common"
+	"go.temporal.io/server/common/headers"
 	"go.temporal.io/server/common/metrics"
+	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
 // Ensure that Activity implements chasm.StateMachine interface
@@ -335,3 +340,449 @@ var TransitionResetAttemptFailedToScheduled = chasm.NewTransition(
 		return a.applyResetAttemptFailedToScheduled(ctx, event)
 	},
 )
+
+// Transition bodies. Each is the body of exactly one chasm.Transition above, which has
+// already validated that the activity is in a legal source state.
+
+// applyScheduled is the body of TransitionScheduled: it opens the first attempt and
+// emits the dispatch, ScheduleToStart and ScheduleToClose tasks.
+func (a *Activity) applyScheduled(ctx chasm.MutableContext) error {
+	attempt := a.LastAttempt.Get(ctx)
+
+	attempt.Count++
+	attempt.Stamp++
+
+	// Start delay defers the dispatch and extends ScheduleToClose and ScheduleToStart timeouts. StartToClose and
+	// Heartbeat timeouts are unaffected as they only start when a worker picks up the task.
+	dispatchTime := a.firstDispatchTime()
+	attempt.DispatchTime = timestamppb.New(dispatchTime)
+
+	if timeout := a.GetScheduleToStartTimeout().AsDuration(); timeout > 0 {
+		ctx.AddTask(
+			a,
+			chasm.TaskAttributes{
+				ScheduledTime: dispatchTime.Add(timeout),
+			},
+			&activitypb.ScheduleToStartTimeoutTask{
+				Stamp: attempt.GetStamp(),
+			})
+	}
+
+	if deadline := a.scheduleToCloseDeadline(); !deadline.IsZero() {
+		a.ScheduleToCloseStamp++
+		ctx.AddTask(
+			a,
+			chasm.TaskAttributes{
+				ScheduledTime: deadline,
+			},
+			&activitypb.ScheduleToCloseTimeoutTask{Stamp: a.GetScheduleToCloseStamp()})
+	}
+
+	dispatchAttrs := chasm.TaskAttributes{}
+	if dispatchTime.After(a.ScheduleTime.AsTime()) {
+		dispatchAttrs.ScheduledTime = dispatchTime
+	}
+	ctx.AddTask(
+		a,
+		dispatchAttrs,
+		a.newActivityDispatchTask(ctx))
+
+	return nil
+}
+
+// applyRescheduled is the body of TransitionRescheduled: it records the failed attempt
+// and re-emits the dispatch and ScheduleToStart tasks at the retry time.
+func (a *Activity) applyRescheduled(ctx chasm.MutableContext, event rescheduleEvent) error {
+	if err := a.applyFailedAttempt(ctx, event); err != nil {
+		return err
+	}
+
+	attempt := a.LastAttempt.Get(ctx)
+	retryScheduledTime := dispatchTimeForRetry(attempt).AsTime()
+	attempt.DispatchTime = timestamppb.New(retryScheduledTime)
+
+	if timeout := a.GetScheduleToStartTimeout().AsDuration(); timeout > 0 {
+		ctx.AddTask(
+			a,
+			chasm.TaskAttributes{
+				ScheduledTime: retryScheduledTime.Add(timeout),
+			},
+			&activitypb.ScheduleToStartTimeoutTask{
+				Stamp: attempt.GetStamp(),
+			})
+	}
+
+	ctx.AddTask(
+		a,
+		chasm.TaskAttributes{
+			ScheduledTime: retryScheduledTime,
+		},
+		a.newActivityDispatchTask(ctx))
+
+	return nil
+}
+
+// applyStarted is the body of TransitionStarted: it records the worker that picked the
+// attempt up and emits the StartToClose and Heartbeat tasks.
+func (a *Activity) applyStarted(ctx chasm.MutableContext, request *historyservice.RecordActivityTaskStartedRequest) error {
+	attempt := a.LastAttempt.Get(ctx)
+	attempt.StartedTime = timestamppb.New(ctx.Now(a))
+	attempt.StartedStamp = request.GetStamp()
+	// Record the first-ever worker pickup time once and never update on retries or resets.
+	if a.FirstAttemptStartedTime == nil {
+		a.FirstAttemptStartedTime = attempt.GetStartedTime()
+	}
+	attempt.StartRequestId = request.GetRequestId()
+	attempt.LastWorkerIdentity = request.GetPollRequest().GetIdentity()
+	attempt.SdkName = ctx.RequestHeader(headers.ClientNameHeaderName)
+	attempt.SdkVersion = ctx.RequestHeader(headers.ClientVersionHeaderName)
+	if versionDirective := request.GetVersionDirective().GetDeploymentVersion(); versionDirective != nil {
+		attempt.LastDeploymentVersion = &deploymentpb.WorkerDeploymentVersion{
+			BuildId:        versionDirective.GetBuildId(),
+			DeploymentName: versionDirective.GetDeploymentName(),
+		}
+	}
+	startTime := attempt.GetStartedTime().AsTime()
+	ctx.AddTask(
+		a,
+		chasm.TaskAttributes{
+			ScheduledTime: startTime.Add(a.GetStartToCloseTimeout().AsDuration()),
+		},
+		&activitypb.StartToCloseTimeoutTask{
+			Stamp: a.LastAttempt.Get(ctx).GetStamp(),
+		})
+
+	if heartbeatTimeout := a.GetHeartbeatTimeout().AsDuration(); heartbeatTimeout > 0 {
+		ctx.AddTask(
+			a,
+			chasm.TaskAttributes{
+				ScheduledTime: startTime.Add(heartbeatTimeout),
+			},
+			&activitypb.HeartbeatTimeoutTask{
+				Stamp: attempt.GetStamp(),
+			})
+	}
+
+	return nil
+}
+
+// applyCompleted is the body of TransitionCompleted: it records the successful outcome.
+func (a *Activity) applyCompleted(ctx chasm.MutableContext, event completeEvent) error {
+	return a.StoreOrSelf(ctx).RecordCompleted(ctx, func(ctx chasm.MutableContext) error {
+		req := event.req.GetCompleteRequest()
+
+		attemptWasStarted := a.hasAttemptInProgress()
+		attempt := a.LastAttempt.Get(ctx)
+		if !attemptWasStarted {
+			// RespondActivityTaskCompletedById can complete an activity when no attempt is in
+			// progress.
+			attempt.StartedTime = timestamppb.New(ctx.Now(a))
+			if a.FirstAttemptStartedTime == nil {
+				a.FirstAttemptStartedTime = attempt.StartedTime
+			}
+		}
+		attempt.CompleteTime = timestamppb.New(ctx.Now(a))
+		attempt.LastWorkerIdentity = req.GetIdentity()
+		outcome := a.Outcome.Get(ctx)
+		outcome.Variant = &activitypb.ActivityOutcome_Successful_{
+			Successful: &activitypb.ActivityOutcome_Successful{
+				Output: req.GetResult(),
+			},
+		}
+
+		a.emitOnCompletedMetrics(ctx, event.baseHandler, event.enrichedHandler, req.GetResult(), attemptWasStarted)
+
+		return nil
+	})
+}
+
+// applyFailed is the body of TransitionFailed: it records a terminal failure outcome.
+// Not to be confused with applyFailedAttempt/recordFailedAttempt, which record the
+// failure of a single attempt that may still be retried.
+func (a *Activity) applyFailed(ctx chasm.MutableContext, event failedEvent) error {
+	return a.StoreOrSelf(ctx).RecordCompleted(ctx, func(ctx chasm.MutableContext) error {
+		req := event.req.GetFailedRequest()
+		a.Outcome.Get(ctx).RetryState = event.retryState
+
+		attempt := a.LastAttempt.Get(ctx)
+		attempt.LastWorkerIdentity = req.GetIdentity()
+
+		// A worker may respond failed without a Failure. Synthesize a generic terminal failure so
+		// the closed activity still exposes a consumable outcome; otherwise PollActivityExecution
+		// returns a nil outcome and a client cannot tell the closed activity apart from one that
+		// simply has no result yet, and polls forever. (Workflow activities tolerate a nil failure
+		// because the SDK wraps the failed history event in an ActivityError; a standalone activity
+		// returns the raw outcome with no such wrapper.)
+		failure := req.GetFailure()
+		if failure == nil {
+			failure = &failurepb.Failure{Message: "activity task failed without failure details"}
+		}
+
+		if err := a.recordFailedAttempt(ctx, 0, activitypb.ACTIVITY_RETRY_INTERVAL_SOURCE_UNSPECIFIED, failure, ctx.Now(a), true); err != nil {
+			return err
+		}
+
+		a.emitOnFailedMetrics(ctx, event.baseHandler, event.enrichedHandler, req.GetFailure())
+
+		return nil
+	})
+}
+
+// applyTerminated is the body of TransitionTerminated: it records the terminated outcome.
+func (a *Activity) applyTerminated(ctx chasm.MutableContext, event terminateEvent) error {
+	return a.StoreOrSelf(ctx).RecordCompleted(ctx, func(ctx chasm.MutableContext) error {
+		a.TerminateState = &activitypb.ActivityTerminateState{
+			RequestId: event.request.RequestID,
+		}
+		outcome := a.Outcome.Get(ctx)
+		failure := &failurepb.Failure{
+			Message: event.request.Reason,
+			FailureInfo: &failurepb.Failure_TerminatedFailureInfo{
+				TerminatedFailureInfo: &failurepb.TerminatedFailureInfo{
+					Identity: event.request.Identity,
+				},
+			},
+		}
+		outcome.Variant = &activitypb.ActivityOutcome_Failed_{
+			Failed: &activitypb.ActivityOutcome_Failed{
+				Failure: failure,
+			},
+		}
+
+		a.emitOnTerminatedMetrics(event.metricsHandler)
+
+		return nil
+	})
+}
+
+// applyCancelRequested is the body of TransitionCancelRequested: it records the cancel
+// request and drops any deferred reset intent.
+func (a *Activity) applyCancelRequested(ctx chasm.MutableContext, req *workflowservice.RequestCancelActivityExecutionRequest) error {
+	a.CancelState = &activitypb.ActivityCancelState{
+		Identity:    req.GetIdentity(),
+		RequestId:   req.GetRequestId(),
+		Reason:      req.GetReason(),
+		RequestTime: timestamppb.New(ctx.Now(a)),
+	}
+	// Cancel takes precedence over a pending reset so clear its deferred intents
+	a.ResetRestoreOptions = false
+	a.ResetShouldClearHeartbeat = false
+
+	return nil
+}
+
+// applyCanceled is the body of TransitionCanceled: it records the canceled outcome.
+func (a *Activity) applyCanceled(ctx chasm.MutableContext, event cancelEvent) error {
+	return a.StoreOrSelf(ctx).RecordCompleted(ctx, func(ctx chasm.MutableContext) error {
+		outcome := a.Outcome.Get(ctx)
+		failure := &failurepb.Failure{
+			Message: "Activity canceled",
+			FailureInfo: &failurepb.Failure_CanceledFailureInfo{
+				CanceledFailureInfo: &failurepb.CanceledFailureInfo{
+					Details:  event.details,
+					Identity: a.GetCancelState().GetIdentity(),
+				},
+			},
+		}
+		outcome.Variant = &activitypb.ActivityOutcome_Failed_{
+			Failed: &activitypb.ActivityOutcome_Failed{
+				Failure: failure,
+			},
+		}
+
+		a.emitOnCanceledMetrics(ctx, event.metricsHandler, event.fromStatus)
+
+		return nil
+	})
+}
+
+// applyTimedOut is the body of TransitionTimedOut: it records the timeout failure for the
+// relevant timeout type and closes the activity.
+func (a *Activity) applyTimedOut(ctx chasm.MutableContext, event timeoutEvent) error {
+	timeoutType := event.timeoutType
+
+	return a.StoreOrSelf(ctx).RecordCompleted(ctx, func(ctx chasm.MutableContext) error {
+		a.Outcome.Get(ctx).RetryState = event.retryState
+		priorAttemptFailure := a.LastAttempt.Get(ctx).GetLastFailureDetails().GetFailure()
+		var err error
+		switch timeoutType {
+		case enumspb.TIMEOUT_TYPE_SCHEDULE_TO_START,
+			enumspb.TIMEOUT_TYPE_SCHEDULE_TO_CLOSE:
+			err = a.recordScheduleToStartOrCloseTimeoutFailure(
+				ctx,
+				timeoutType,
+				fmt.Sprintf(common.FailureReasonActivityTimeout, timeoutType.String()),
+				priorAttemptFailure,
+			)
+		case enumspb.TIMEOUT_TYPE_START_TO_CLOSE:
+			failure := createStartToCloseTimeoutFailure()
+			failure.GetTimeoutFailureInfo().LastHeartbeatDetails = a.lastHeartbeatDetails(ctx)
+			failure.Cause = priorAttemptFailure
+			err = a.recordFailedAttempt(ctx, 0, activitypb.ACTIVITY_RETRY_INTERVAL_SOURCE_UNSPECIFIED, failure, ctx.Now(a), true)
+		case enumspb.TIMEOUT_TYPE_HEARTBEAT:
+			failure := createHeartbeatTimeoutFailure()
+			failure.GetTimeoutFailureInfo().LastHeartbeatDetails = a.lastHeartbeatDetails(ctx)
+			failure.Cause = priorAttemptFailure
+			err = a.recordFailedAttempt(ctx, 0, activitypb.ACTIVITY_RETRY_INTERVAL_SOURCE_UNSPECIFIED, failure, ctx.Now(a), true)
+		default:
+			err = fmt.Errorf("unhandled activity timeout: %v", timeoutType)
+		}
+		if err != nil {
+			return err
+		}
+
+		retryPreventedByScheduleToClose := event.retryState == enumspb.RETRY_STATE_TIMEOUT &&
+			(timeoutType == enumspb.TIMEOUT_TYPE_START_TO_CLOSE ||
+				timeoutType == enumspb.TIMEOUT_TYPE_HEARTBEAT)
+		if retryPreventedByScheduleToClose {
+			if err := a.recordScheduleToStartOrCloseTimeoutFailure(
+				ctx,
+				enumspb.TIMEOUT_TYPE_SCHEDULE_TO_CLOSE,
+				common.FailureReasonActivityRetryScheduleToCloseTimeout,
+				priorAttemptFailure,
+			); err != nil {
+				return err
+			}
+		}
+
+		a.emitOnTimedOutMetrics(event.metricsHandler, timeoutType)
+
+		return nil
+	})
+}
+
+// applyPaused is the body of TransitionPaused: it records the pause and bumps the attempt
+// stamp so that any pending dispatch task is invalidated and the activity is not
+// dispatched while paused.
+func (a *Activity) applyPaused(ctx chasm.MutableContext, event pauseEvent) error {
+	a.recordPauseState(ctx, event)
+	attempt := a.LastAttempt.Get(ctx)
+	attempt.Stamp++
+	return nil
+}
+
+// applyPauseRequested is the body of TransitionPauseRequested: it records the pause
+// request. Unlike applyPaused it does not bump the attempt stamp, because the worker
+// remains in charge of the attempt and its StartToClose and Heartbeat tasks must stay
+// valid.
+func (a *Activity) applyPauseRequested(ctx chasm.MutableContext, event pauseEvent) error {
+	a.recordPauseState(ctx, event)
+	return nil
+}
+
+// applyUnpaused is the body of TransitionUnpaused: it bumps the attempt stamp and re-emits
+// the dispatch and ScheduleToStart tasks so that another attempt is made.
+func (a *Activity) applyUnpaused(
+	ctx chasm.MutableContext,
+	event unpauseEvent,
+) error {
+	attempt := a.LastAttempt.Get(ctx)
+	dispatchTime := a.unpauseDispatchTime(ctx, event)
+	attempt.DispatchTime = timestamppb.New(dispatchTime)
+
+	attempt.Stamp++
+	if timeout := a.GetScheduleToStartTimeout().AsDuration(); timeout > 0 {
+		ctx.AddTask(
+			a,
+			chasm.TaskAttributes{ScheduledTime: dispatchTime.Add(timeout)},
+			&activitypb.ScheduleToStartTimeoutTask{Stamp: attempt.GetStamp()})
+	}
+	ctx.AddTask(
+		a,
+		chasm.TaskAttributes{ScheduledTime: dispatchTime},
+		a.newActivityDispatchTask(ctx))
+	return nil
+}
+
+// applyReset is the body of TransitionReset: it returns the activity to attempt 1, bumps the
+// attempt stamp to invalidate any pending dispatch task, and re-emits the dispatch and
+// ScheduleToStart tasks at the reset time.
+func (a *Activity) applyReset(ctx chasm.MutableContext, event resetEvent) error {
+	attempt := a.LastAttempt.Get(ctx)
+	attempt.Count = 1
+	attempt.Stamp++
+	attempt.CurrentRetryInterval = nil
+	attempt.CurrentRetryIntervalSource = activitypb.ACTIVITY_RETRY_INTERVAL_SOURCE_UNSPECIFIED
+	if event.req.GetResetHeartbeat() {
+		a.clearHeartbeatDetails(ctx)
+	}
+	dispatchTime := a.dispatchTimeRespectingStartDelay(event.resetTime)
+	attempt.DispatchTime = timestamppb.New(dispatchTime)
+	if timeout := a.GetScheduleToStartTimeout().AsDuration(); timeout > 0 {
+		ctx.AddTask(
+			a,
+			chasm.TaskAttributes{ScheduledTime: dispatchTime.Add(timeout)},
+			&activitypb.ScheduleToStartTimeoutTask{Stamp: attempt.GetStamp()},
+		)
+	}
+	ctx.AddTask(
+		a,
+		chasm.TaskAttributes{ScheduledTime: dispatchTime},
+		a.newActivityDispatchTask(ctx),
+	)
+	a.emitOnResetMetrics(event.metricsHandler)
+	return nil
+}
+
+// applyResetAttemptFailedToPaused is the body of
+// TransitionResetAttemptFailedToPaused: it applies the deferred reset and leaves the
+// activity paused, emitting no dispatch task.
+func (a *Activity) applyResetAttemptFailedToPaused(ctx chasm.MutableContext, event rescheduleEvent) error {
+	attempt := a.LastAttempt.Get(ctx)
+	a.ResetShouldPause = false
+	a.applyDeferredOptionRestore(ctx)
+	a.applyDeferredHeartbeatClear(ctx)
+	attempt.Count = 1
+	attempt.Stamp++
+	if err := a.recordFailedAttempt(ctx, event.retryInterval, event.retryIntervalSource, event.failure, ctx.Now(a), false); err != nil {
+		return err
+	}
+	// Reset discards the retry backoff
+	attempt.CurrentRetryInterval = nil
+	attempt.CurrentRetryIntervalSource = activitypb.ACTIVITY_RETRY_INTERVAL_SOURCE_UNSPECIFIED
+	attempt.DispatchTime = nil
+	return nil
+}
+
+// applyResetAttemptFailedToScheduled is the body of
+// TransitionResetAttemptFailedToScheduled: it applies the deferred reset and re-emits
+// the dispatch and ScheduleToStart tasks.
+func (a *Activity) applyResetAttemptFailedToScheduled(ctx chasm.MutableContext, event rescheduleEvent) error {
+	attempt := a.LastAttempt.Get(ctx)
+	currentTime := ctx.Now(a)
+
+	a.ResetShouldPause = false
+	a.applyDeferredOptionRestore(ctx)
+	a.applyDeferredHeartbeatClear(ctx)
+
+	attempt.Count = 1
+	attempt.Stamp++
+
+	if err := a.recordFailedAttempt(ctx, event.retryInterval, event.retryIntervalSource, event.failure, currentTime, false); err != nil {
+		return err
+	}
+	// Reset discards the retry backoff
+	attempt.CurrentRetryInterval = nil
+	attempt.CurrentRetryIntervalSource = activitypb.ACTIVITY_RETRY_INTERVAL_SOURCE_UNSPECIFIED
+
+	dispatchTime := a.dispatchTimeRespectingStartDelay(currentTime)
+	attempt.DispatchTime = timestamppb.New(dispatchTime)
+	if timeout := a.GetScheduleToStartTimeout().AsDuration(); timeout > 0 {
+		ctx.AddTask(
+			a,
+			chasm.TaskAttributes{
+				ScheduledTime: dispatchTime.Add(timeout),
+			},
+			&activitypb.ScheduleToStartTimeoutTask{
+				Stamp: attempt.GetStamp(),
+			})
+	}
+	ctx.AddTask(
+		a,
+		chasm.TaskAttributes{
+			ScheduledTime: dispatchTime,
+		},
+		a.newActivityDispatchTask(ctx))
+
+	return nil
+}
